requirements.txt
# Core Dependencies
PyQt6>=6.4.0
requests>=2.31.0
python-dotenv>=1.0.0
filelock>=3.12.0
PyPDF2>=3.0.0
numpy>=1.24.0
opencv-python-headless>=4.7.0
pyautogui>=0.9.54
Pillow>=10.0.0

# Google Generative AI
google-genai

# Development Dependencies
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
streamlit>=1.30.0
